Better CrusteeTable.get() perfs
Refactored to use ArrayLists explicitely the inner Tables class.
This avoids polymorphic list calls and allows to use a for+get loop instead
of iterator-style looping.

Also, SSTableReader.get() no longer returns an Optional. While this is
usually a better practice than returning null, we should avoid any extra
object in memory, and Optional's methods such as map() do many  extra
operations internally, which could be costly considering SSTableReader is on the
critical path.

This should be tested in a benchmark obviously, but I think this simpler
design should lead to better performances, at the expense of a slightly
more verbose code.